improve evaluation storage + fix minor bugs

+++ URingPaxos/clients/SMR/src/java/ch/usi/da/smr/Client.java
@@ -227,7 +227,7 @@ public class Client implements Receiver {
 			Thread t = new Thread("Command Sender " + i) {
 				@Override
 				public void run() {
-					allLatencies.put(System.currentTimeMillis(), 0);
+					allLatencies.put(System.currentTimeMillis(), 0L);
 					int sendCount = 0;
 					while (sendCount < sendsPerThread) {
 						int id = send_id.incrementAndGet();
@@ -268,7 +268,7 @@ public class Client implements Receiver {
 
 				private void sleepForThinkingTime() {
 					try {
-						int timeToSleep = (Math.random() * 10000.0) % thinkingTime + 5;
+						int timeToSleep = ((int) (Math.random() * 10000.0)) % thinkingTime + 5;
 						Thread.sleep(timeToSleep);
 					} catch (Exception ex) {
 						logger.error("we have a problem", ex);
